+++ banned-import/main/kotlin/com/javiersc/detekt/rules/banned/import/BannedImport.kt
@@ -1,4 +1,4 @@
-package com.javiersc.detektRules.bannedImport
+package com.javiersc.detekt.rules.banned.import
 
 import io.gitlab.arturbosch.detekt.api.CodeSmell
 import io.gitlab.arturbosch.detekt.api.Config