+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+- package to `com.javiersc.detekt.rules.banned.import`
+
 ### Deprecated
 
 ### Removed
@@ -14,10 +16,10 @@
 
 ### Updated
 
+- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.33`
 - `gradle -> 7.3.3`
 - `io.kotest:kotest-assertions-core -> 5.0.3`
 - `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.6.10`
-- `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.24`
 - `io.gitlab.arturbosch.detekt:detekt-test -> 1.19.0`
 - `io.gitlab.arturbosch.detekt:detekt-api -> 1.19.0`